requirements-gpu.txt
tensorflow-gpu
sklearn
git-lfs
pylint
pytest
travis
pandas
dataclasses
numpy
matplotlib
tqdm
pandas
fire
opencv-python